x-automation/requirements.txt
tweepy
python-dotenv
requests
schedule
beautifulsoup4
feedparser
